% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_EloCalc_season.R
\name{AN_EloCalc_season}
\alias{AN_EloCalc_season}
\title{AN_EloCalc_season}
\usage{
AN_EloCalc_season(data.m, k = 35, a = 1/1400, factors = c(0.05, -0.6))
}
\arguments{
\item{data.m}{matrix (!) that should have columns "Res","RowPrev","Q0","Q1", "HWpct", "Dpct", "AWpct"
"RowPrev" should be equal to row_number() if it is the first occurrence
"Q0" should be filled for the first occurrence per team
"Q1" should be equal to Q0
nrow(data.m) should be even, odd rows for home teams, even rows for away teams}

\item{k}{k-factor}

\item{a}{division factor}

\item{startingQs}{tibble with columns "Team" and "Q0"}

\item{b1}{home team correction factor}

\item{b2}{away team correction factor}
}
\value{
tibble with columns "Div","Season","Date","HomeTeam","AwayTeam","Res","HWpct","Dpct","AWpct","QH0","QH1","QA0","QA1"
}
\description{
wrapper function that downloads multiple seasons for multiple league from one of the source websites
}
