% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_EloMatrix_season.R
\name{AN_EloMatrix_season}
\alias{AN_EloMatrix_season}
\title{AN_EloMatrix_season}
\usage{
AN_EloMatrix_season(data, startingQs)
}
\arguments{
\item{data}{match data tibble that should contain one season for one league with columns "Div","Season","Date","HomeTeam","AwayTeam" and "Res"}

\item{startingQs}{tibble with columns "Team" and "Q0"}
}
\value{
a matrix with columns
}
\description{
converts a match data tibble into a matrix for faster optimalisation
}
