% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_EloStartingQ_season.R
\name{AN_EloStartingQ_season}
\alias{AN_EloStartingQ_season}
\title{AN_EloStartingQ_season}
\usage{
AN_EloStartingQ_season(data, level1.baseQ = 2000, level1.spread = 500,
  level1.maxQ = Inf)
}
\arguments{
\item{data}{match data tibble that should contain one season for one country with columns "Div","Season","Date","HomeTeam","AwayTeam" and "Res"}

\item{level1.baseQ}{the mean Q at the highest level}

\item{level1.spread}{the maximum Q delta between the best and worst team at the highest level}

\item{level1.maxQ}{the maximum Q of the best team at the highest level}
}
\value{
a tibble with columns "Team" and "Q0"
}
\description{
provides starting Qs for multiple levels for a country
}
