% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_AddSeason.R
\name{CL_AddSeason}
\alias{CL_AddSeason}
\title{CL_AddSeason}
\usage{
CL_AddSeason(data)
}
\arguments{
\item{data}{match data tibble that should contain columns Div, Season, Date, HomeTeam, AwayTeam}
}
\value{
same as data input but with Season filled in
}
\description{
determine the seasons for the entire data frame and fill the Season column
}
