% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_BetExplorer_data.R
\name{DL_BetExplorer_data}
\alias{DL_BetExplorer_data}
\title{DL_BetExplorer_data}
\usage{
DL_BetExplorer_data(divs, year1 = (as.numeric(format(Sys.Date(), "\%Y")) - 1))
}
\value{
tibble with columns Div (e.g. "NLD1"),"Season",Year (e.g. 2005) and URL
}
\description{
creates tibble with URLs per league per season
}
