% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_BetExplorer_season.R
\name{DL_BetExplorer_season}
\alias{DL_BetExplorer_season}
\title{DL_BetExplorer_season}
\usage{
DL_BetExplorer_season(div, url)
}
\arguments{
\item{div}{e.g. "DEU2"}

\item{url}{e.g. "http://www.betexplorer.com/soccer/germany/2-bundesliga-2015-2016/results"}
}
\value{
tibble for one season and one league with all matches, results and bookmaker odds
}
\description{
download one season for one league from www.betexplorer.com
}
