% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_FootballDataCoUk.R
\name{DL_FootballDataCoUk}
\alias{DL_FootballDataCoUk}
\title{DL_FootballDataCoUk}
\usage{
DL_FootballDataCoUk(divs, season1, season2,
  DL.input = DL_FootballDataCoUk_data())
}
\arguments{
\item{divs}{e.g. c("ENG1","FRA")}

\item{season1}{e.g. ("2001-2002")}

\item{season2}{e.g. ("2005-2006")}

\item{DL.input}{usually output of DL_FootballDataCoUk_data()}
}
\value{
tibble for multiple seasons and leagues with all matches, results and bookmaker odds
}
\description{
download multiple seasons for multiple league from www.football-data.co.uk
}
