% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_FootballDataCoUk_data.R
\name{DL_FootballDataCoUk_data}
\alias{DL_FootballDataCoUk_data}
\title{DL_FootballDataCoUk_data}
\usage{
DL_FootballDataCoUk_data()
}
\value{
tibble with all matches
}
\description{
download multiple seasons for multiple league from www.football-data.co.uk
}
