% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_FootballDataCoUk_season.R
\name{DL_FootballDataCoUk_season}
\alias{DL_FootballDataCoUk_season}
\title{DL_FootballDataCoUk_season}
\usage{
DL_FootballDataCoUk_season(div, url)
}
\arguments{
\item{div}{e.g. ENG0}

\item{url}{e.g. http://www.football-data.co.uk/mmz4281/1516/E0.csv}
}
\value{
tibble with all matches
}
\description{
download one season for one league from www.football-data.co.uk
}
