% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_data.R
\name{DL_Soccerway_data}
\alias{DL_Soccerway_data}
\title{DL_Soccerway_data}
\usage{
DL_Soccerway_data(divs = "DEU3", year1 = (as.numeric(format(Sys.Date(),
  "\%Y")) - 1))
}
\value{
tibble with columns Div (e.g. "NLD1"),"Season",Year (e.g. 2005) URL.geo, URL.logo, URL.stadium
}
\description{
creates tibble with the string that is needed to plot correctly
}
