% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_download_logo.R
\name{DL_Soccerway_download_logo}
\alias{DL_Soccerway_download_logo}
\title{DL_Soccerway_download_logo}
\usage{
DL_Soccerway_download_logo(DL.input, subfolder = "~/Desktop/Logos",
  overwrite = FALSE)
}
\arguments{
\item{DL.input}{a tibble that should contain columns FileName (string, with extension) and LogoURL.}

\item{subfolder}{the name of the subfolder where to store thelogos, e.g. "~/Desktop/Logos"}

\item{overwrite}{if TRUE then download all logos and overwrite existing logos in the folder with the same name}
}
\description{
Download logos to the specified folder, which is a subfolder of the working directory.
If the folder doesn't exist then create it.
}
\examples{
setwd("/Users/tristanbains/Desktop/R/rFootballAnalysis");lg="RUS2";
lgs=substr(lg,1,3);df = DL_Soccerway_data(lg); df.logos = DL_Soccerway_season_logo(lgs,df$URL.logo);
DL.input = df.logos \%>\% mutate(FileName=paste0(CountryCode," - ",Team,".png"));DL_Soccerway_download_logo(DL.input)



}
