% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_season_champion.R
\name{DL_Soccerway_season_champion}
\alias{DL_Soccerway_season_champion}
\title{DL_Soccerway_season_champion}
\usage{
DL_Soccerway_season_champion(div, url)
}
\value{
tibble
}
\description{
creates tibble
}
