% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_season_geo.R
\name{DL_Soccerway_season_geo}
\alias{DL_Soccerway_season_geo}
\title{DL_Soccerway_season_geo}
\usage{
DL_Soccerway_season_geo(countryCode, url)
}
\arguments{
\item{countryCode}{e.g. "ESP"}

\item{url}{e.g. "http://nl.soccerway.com/national/spain/segunda-division/20152016/regular-season/r32028/map"}
}
\value{
tibble with columns CountryCode ("ESP"), Team ("UD Almería"), Lat, Long
}
\description{
creates tibble with lat-long pairs per club. url input usually comes from DL_Soccerway_season_data(), the URL.geo column
df = DL_Soccerway_data("NLD1"); DL_Soccerway_season_geo("NLD",df$URL.geo)
}
