% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_season_stadium.R
\name{DL_Soccerway_season_stadium}
\alias{DL_Soccerway_season_stadium}
\title{DL_Soccerway_season_stadium}
\usage{
DL_Soccerway_season_stadium(div, url)
}
\value{
tibble
}
\description{
creates tibble
}
