% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_MapData.R
\name{PL_MapData}
\alias{PL_MapData}
\title{PL_MapData}
\usage{
PL_MapData()
}
\value{
tibble with columns "CountryCode","MapSource","MapString","xmin","xmax","ymin","ymax"
}
\description{
creates tibble with the string that is needed to plot correctly
}
\details{
#@export
}
