% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Ranking.R
\name{ST_Ranking}
\alias{ST_Ranking}
\title{ST_Ranking}
\usage{
ST_Ranking(data, ordering = ST_RankingOrdering())
}
\arguments{
\item{data}{match data tibble that should contain columns "Div","Season","Date","HomeTeam","AwayTeam","FTHG","FTAG" and "Res"}
}
\value{
tibble witH columns "Div","Season","Rank","Date","Team","M","W","D","L","Pts","GF","GA","GD","HM","HW","HD","HL","HPts","HGF","HGA","HGD","AM","AW","AD","AL","APts","AGF","AGA","AGD"
}
\description{
creates season rankings
}
