% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_RankingOrdering.R
\name{ST_RankingOrdering}
\alias{ST_RankingOrdering}
\title{ST_RankingOrdering}
\usage{
ST_RankingOrdering()
}
\arguments{
\item{data}{match data tibble that should contain one season for one league with columns "Div","Season","Date","HomeTeam","AwayTeam" and "Res"}
}
\value{
tibble with columns "Div","Season","Parameter","Order"
}
\description{
creates tibble with ranking criteria per league
}
