% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_Soccerway_season_champion.R
\name{DL_Soccerway_season_champion}
\alias{DL_Soccerway_season_champion}
\title{DL_Soccerway_season_champion}
\usage{
DL_Soccerway_season_champion(div, url)
}
\arguments{
\item{div}{e.g. "NLD1"}

\item{url}{e.g. "http://nl.soccerway.com/national/netherlands/eredivisie/c1/archive"}
}
\value{
tibble with columns Div ("ESP1"), Season, Champion ("Barcelona"), RunnerUp ("Real Madrid")
}
\description{
creates tibble with all the champions in a league. url input usually comes from DL_Soccerway_season_data(), the URL.archive column
df = DL_Soccerway_data("NLD1"); DL_Soccerway_season_champion("NLD",df$URL.archive)
}
