% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_Ranking.R
\name{ST_Ranking}
\alias{ST_Ranking}
\title{ST_Ranking}
\usage{
ST_Ranking(data, ordering = ST_RankingOrdering())
}
\arguments{
\item{data}{match data tibble that should contain one season for one league with columns "Div","Season","Date","HomeTeam","AwayTeam" and "Res"}
}
\value{
tibble with columns "Div","Season","Date","HomeTeam","AwayTeam","Res","HWpct","Dpct","AWpct","QH0","QH1","QA0","QA1"
}
\description{
creates season rankings
}
