% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_RankingOrdering.R
\name{ST_RankingOrdering}
\alias{ST_RankingOrdering}
\title{ST_RankingOrdering}
\usage{
ST_RankingOrdering()
}
\value{
tibble with columns "Div","Season","Parameter","Order"
}
\description{
creates tibble with ranking criteria per league
}
